+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - not equal operator for where method
 - Table::except() method to exclude some data
 - Table::groupThenApply() method to apply aggregate functions
+- Arr::from() method for creating new Arr from a string or array-like object;
 
 ## 0.0.7 - 2022-04-25
 - Review where method, where(price, '>', 100)